[DF-7621] -> [DF-7628] Saving landscape to texture reworked (not rendering yet)
Former-commit-id: d7a5c86389b730fe2e2ec235eb3a4dc8c06a2079

+++ Sources/Internal/Render/Renderer.cpp
@@ -100,8 +100,8 @@ void Uninitialize()
 
 void Reset(const rhi::ResetParam & params)
 {
-    framebufferWidth = params.width * params.scaleX;
-    framebufferHeight = params.height * params.scaleY;
+    framebufferWidth = static_cast<int>(params.width * params.scaleX);
+    framebufferHeight = static_cast<int>(params.height * params.scaleY);
 
     rhi::Reset(params);
 }
